[ES|QL] Dashboard variables follow ups (#208338)
(cherry picked from commit fa7c477ff19855d32f5d8dd348a6b0ce50381c39)

# Conflicts:
#	src/platform/packages/shared/kbn-esql-utils/src/utils/append_to_query.ts
