+++ src/platform/packages/private/kbn-esql-editor/src/esql_editor.tsx
@@ -38,9 +38,12 @@ import memoize from 'lodash/memoize';
 import React, { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { css } from '@emotion/react';
-import { ESQLRealField, ESQLControlVariable } from '@kbn/esql-validation-autocomplete';
+import {
+  type ESQLRealField,
+  ESQLVariableType,
+  type ESQLControlVariable,
+} from '@kbn/esql-validation-autocomplete';
 import { FieldType } from '@kbn/esql-validation-autocomplete/src/definitions/types';
-import { ESQLVariableType } from '@kbn/esql-validation-autocomplete';
 import { EditorFooter } from './editor_footer';
 import { fetchFieldsFromESQL } from './fetch_fields_from_esql';
 import {
