+++ src/platform/packages/shared/kbn-esql-utils/index.ts
@@ -18,6 +18,7 @@ export {
   getESQLWithSafeLimit,
   appendToESQLQuery,
   appendWhereClauseToESQLQuery,
+  appendStatsByToQuery,
   getESQLQueryColumns,
   getESQLQueryColumnsRaw,
   getESQLResults,
@@ -36,6 +37,7 @@ export {
   TextBasedLanguages,
   queryCannotBeSampled,
   mapVariableToColumn,
+  getValuesFromQueryField,
 } from './src';
 
 export { ENABLE_ESQL, FEEDBACK_LINK } from './constants';
