+++ src/platform/packages/shared/kbn-esql-utils/src/utils/query_parsing_helpers.ts
@@ -180,3 +180,20 @@ export const mapVariableToColumn = (
   });
   return columns;
 };
+
+export const getValuesFromQueryField = (queryString: string) => {
+  const validQuery = `${queryString} ""`;
+  const { root } = parse(validQuery);
+  const lastCommand = root.commands[root.commands.length - 1];
+  const columns: ESQLColumn[] = [];
+
+  walk(lastCommand, {
+    visitColumn: (node) => columns.push(node),
+  });
+
+  const column = Walker.match(lastCommand, { type: 'column' });
+
+  if (column) {
+    return `${column.name}`;
+  }
+};
