+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/index.ts
@@ -9,7 +9,6 @@
 
 export type { SuggestionRawDefinition, ItemKind } from './src/autocomplete/types';
 export { ESQLVariableType, type ESQLControlVariable } from './src/shared/types';
-export { inKnownTimeInterval } from './src/shared/helpers';
 export type { CodeAction } from './src/code_actions/types';
 export type {
   FunctionDefinition,
@@ -68,6 +67,7 @@ export {
   isSingleItem,
 } from './src/shared/helpers';
 export { ENRICH_MODES } from './src/definitions/settings';
+export { timeUnits } from './src/definitions/literals';
 export { getFunctionSignatures } from './src/definitions/helpers';
 
 export {
